Refactored files to organise scripts into directories
