--- bin/resources/config.json
@@ -1,9 +1,0 @@
-{
-    "001_REFERENCE_PROJ_ID": "project-GXZ0qvj4kbfjZ2fKpKZbxy8q",
-    "003_DEV_CLINVAR_UPDATE_PROJ_ID": "project-GXXb2K04FK2bGyvV5GVxpYFf",
-    "SLACK_CHANNEL": "egg-test",
-    "TSO500_CONFIG_REPO": "https://github.com/PolarisAstrum-NHS/dummy_vep_helios_config",
-    "TWE_CONFIG_REPO": "https://github.com/PolarisAstrum-NHS/dummy_vep_TWE_config",
-    "CEN_CONFIG_REPO": "https://github.com/PolarisAstrum-NHS/dummy_vep_CEN_config",
-    "TSO500_WORKFLOW_REPO": "https://github.com/PolarisAstrum-NHS/dummy_TSO500_reports_workflow"
-}
