+++ bin/tests/context.py
@@ -2,18 +2,18 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
 
-import annotation_update
-import compare_annotation
-import deployer
-import get_clinvar_files
-import git_handler
-import inspect_vep_logs
-import inspect_workflow_logs
-import login_handler
-import make_vep_test_configs
-import progress_tracker
-import slack_handler
-import utils
-import vep_config_update
-import vep_testing
-import workflow_handler
+import annotation.annotation_update as annotation_update
+import annotation.compare_annotation as compare_annotation
+import util.deployer as deployer
+import annotation.get_clinvar_files as get_clinvar_files
+import util.git_handler as git_handler
+import util.inspect_vep_logs as inspect_vep_logs
+import workflow.inspect_workflow_logs as inspect_workflow_logs
+import util.login_handler as login_handler
+import annotation.make_vep_test_configs as make_vep_test_configs
+import util.progress_tracker as progress_tracker
+import util.slack_handler as slack_handler
+import util.utils as utils
+import vep_config.vep_config_update as vep_config_update
+import util.vep_testing as vep_testing
+import workflow.workflow_handler as workflow_handler
