+++ bin/vep_config/vep_config_update.py
@@ -10,17 +10,17 @@ import re
 import glob
 import sys
 
-import vep_testing
-import deployer
-from login_handler import LoginHandler
-from slack_handler import SlackHandler
-from utils import (get_prod_version,
-                   load_config,
-                   load_config_repo,
-                   check_proj_folder_exists)
-from git_handler import GitHandler
-import utils
-from progress_tracker import VepProgressTracker as Tracker
+import util.vep_testing as vep_testing
+import util.deployer as deployer
+from util.login_handler import LoginHandler
+from util.slack_handler import SlackHandler
+from util.utils import (get_prod_version,
+                        load_config,
+                        load_config_repo,
+                        check_proj_folder_exists)
+from util.git_handler import GitHandler
+import util.utils as utils
+from util.progress_tracker import VepProgressTracker as Tracker
 
 logger = logging.getLogger("main log")
